refactor(settings):change name component temp
